+++ courses/course_factory.go
@@ -5,7 +5,20 @@ import (
 )
 
 /**
- * The [NewCourse] function...
+ * The [NewCourses] function creates new Course instances from slice of raw maps.
+ */
+func NewCourses (rawCourses []map[string]interface{}) []Course {
+  newCourses := make ([]Course, len (rawCourses))
+
+  for i, rawCourse := range rawCourses {
+    newCourses[i] = NewCourse (rawCourse)
+  }
+
+  return newCourses
+}
+
+/**
+ * The [NewCourse] function creates a new Course instance from a raw map.
  */
 func NewCourse (rawCourse map[string]interface{}) Course {
   created, _ := time.Parse (time.RFC3339, rawCourse["created"].(string))